apps/llm-pipeline/pyproject.toml
[project]
name = "bushiri-llm-pipeline"
version = "0.1.0"
description = "PydanticAI pipeline for BUSHIRI market post parsing"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.124.0",
  "pydantic>=2.12.0",
  "pydantic-ai-slim[google]>=1.31.0",
  "uvicorn[standard]>=0.38.0"
]

[project.optional-dependencies]
test = []

[tool.uv]
package = false
